Abstract codegen code into ColumnarBatchScan
